refactor AppDeniedMetric into throttle_metric_result.go
